#11 set dependency versions to avoid installation errors
The previous client dependency version of "k8s-dogu-operator" was unset. With
version 0.16.0 the custom deployment mechanism was removed. Earlier versions of
the operator will produce an error while trying to create a deployment.

Also, the client dependency version of "cesapp" was set to an impossible
version in order to fail early as possible. This is because the nginx-ingress
dogu is architecturally unable to be run inside a cesapp driven CES instance.

"ces-setup" is also incapable of working with this dogu thus an impossible
version was likewise set.