#17 enable gzip compression
It has been concluded, that dogus should not send gzipped responses.
Instead, the nginx-ingress should compress responses.
Gzipped responses will be disabled in the individual dogus.
Additionally, in the k8s-service-discovery the Ingress objects will be written in such a way that the nginx-ingress sets the `Accept-Encoding` header to `identity`.