[#45] Enable Snippet annotations
Snippet annotations are needed to use redirect of the k8s-service-discovery as the annotation: `nginx.ingress.kubernetes.io/permanent-redirect` only works for plain hostnames. If the want to preserve the request path the annotation `nginx.ingress.kubernetes.io/server-snippet` needs to be used. The server-snippet annotation needs annotations-risk-level config to be set to `Critical`.